+++ src/nginx_module/access_phase_handler.c
@@ -2,6 +2,7 @@
 #include <ngx_http.h>
 #include "access_phase_handler.h"
 #include "filter_context.h"
+#include "firetail_config.h"
 #include "firetail_module.h"
 #include <json-c/json.h>
