+++ src/nginx_module/filter_response_body.c
@@ -3,6 +3,7 @@
 #include <json-c/json.h>
 #include "filter_context.h"
 #include "filter_response_body.h"
+#include "firetail_config.h"
 #include "firetail_module.h"
 
 struct ValidateResponseBody_return {
